feat: implement mission site management and update dependencies in game state
prompt:
Change the #sym:GameState  so that instead of `deployedMissionIds` it has `deployedMissionSiteIds`. To do that, you must introduce a new type, `MissionSite`. It will have a reference to the Mission it originated from in key `missionId`, but it will also have its own `id` and also the list of agent IDs deployed to it as well as a key of type `MissionSiteState` that can have 3 values: `Active`, `Successful` and `Failed`. When agents are deployed a mission site is in "Active" state. After one turn it becomes "Successful" if at least 2 agents were deployed to it, "Failed" otherwise.

The mission site ID should be just a globally incrementing counter, similar to `nextAgentId`.

Also update #file:LeadCards.tsx  display so that if a lead depends on a mission, it means that the `deployedMissionSiteIds` must have a mission that has such mission as its `missionId` and its status is `Successful`.